+++ VerticalWarehouses/Machine Automation Service/Ferretto.VW.MAS.AutomationService/HostedServices/NotificationTelemetryService.cs
@@ -131,7 +131,15 @@ namespace Ferretto.VW.MAS.AutomationService
                             var dataLayer = this.ServiceScopeFactory.CreateScope().ServiceProvider.GetRequiredService<IDataLayerService>();
                             var machine = machineDataProvider.Get();
                             // save the database to server
-                            dataLayer.CopyMachineDatabaseToServer(machine.BackupServer, machine.BackupServerUsername, machine.BackupServerPassword, machineDataProvider.GetSecondaryDatabase(), machine.SerialNumber);
+                            try
+                            {
+                                dataLayer.CopyMachineDatabaseToServer(machine.BackupServer, machine.BackupServerUsername, machineDataProvider.GetBackupServerPassword(), machineDataProvider.GetSecondaryDatabase(), machine.SerialNumber);
+                            }
+                            catch (ApplicationException ex)
+                            {
+                                var errorProvider = scope.ServiceProvider.GetRequiredService<IErrorsProvider>();
+                                errorProvider.RecordNew(DataModels.MachineErrorCode.BackupDatabaseOnServer, additionalText: ex.Message);
+                            }
                         }
                     }
                     catch (Exception ex)